requirements.txt
streamlit==1.32.0
pypdf==4.1.0
groq>=0.9.0
tavily-python==0.3.3
pandas==2.2.1
httpx<0.27
